stub: add requests.sessions module
